# perf: skip shifting on bitslices