Some bitslice manipulation can be skipped once we know we are looking for a bigger gap.

Before:
```
**read_heavy** with prefill_fraction 0.95
ValueStorage:
1572864 operations across 1 thread(s) in 875.30225ms; time/op = 556ns

**insert_heavy** with prefill_fraction 0.0
ValueStorage:
1572864 operations across 1 thread(s) in 1.914102125s; time/op = 1.216µs

**update_heavy** with prefill_fraction 0.5
ValueStorage:
1572864 operations across 1 thread(s) in 1.640245875s; time/op = 1.042µs
```

After:
```
**read_heavy** with prefill_fraction 0.95
ValueStorage:
1572864 operations across 1 thread(s) in 899.85575ms; time/op = 572ns

**insert_heavy** with prefill_fraction 0.0
ValueStorage:
1572864 operations across 1 thread(s) in 1.8158965s; time/op = 1.154µs

**update_heavy** with prefill_fraction 0.5
ValueStorage:
1572864 operations across 1 thread(s) in 1.512245041s; time/op = 961ns
```

### Note
This has not proved enough performance advantage (change is within 5% improvement) to justify the increase in complexity, since we expected the improvement to be more drastic.

However, we'll keep it as a pull request so we can pick it up in case we find how to leverage this approach better.